Simple trick system
Uses two separate modes which will be selected in the main menu. A hard or easy mode. The easy mode is all animated, the hard mode uses no animations and relies on the player to make sure the car lands up-right - could implement further shit or just scrap the mode system I was just bored