Updated gameover screen
Alternating game over text semi bundily as per the teachers requests.  Could make it say random shit but meh.. Added functionality of the back button in gameover screen to restart the scene and yeh. Noticed there is a bug with pausing, it's not even coded in anymore I don't think.. it's a few lines but i'm tired as fuck atm, bool is m_bIsPaused or some shiet in CH